refactor: move the legislature domain out of the uslm module
AmendingAction, UscReference, BillAmendment, and BillDiff described
legislative behaviour from inside a module named after a markup language.
An amendment strikes, inserts, or redesignates whether the bill arrived as
USLM XML, as a scanned PDF, or from a state that publishes in some other
format. Confusing the two mattered as soon as a second format appeared: a
reader for a state code needs these types and has no USLM, and a court
opinion needs none of them.

They now live in src/legislature. src/uslm keeps what is genuinely about the
format: the element tree, source credits, text content fields, and parsing.

Pure move. 201 passed before, 201 after.

One coupling survives: AmendingAction::from_str returns USLMError, so the
domain module still reaches into the format module for an error type. That
wants a legislature error of its own.

Part of #51
